REFACTORED PREFABS | Prefab instances now remember changes made
When changes are made to a prefab instance's component, that component will no longer be updated by changes made to the original prefab. It will also be highlighted in the inspector. Prefab editor still needs some changes to work correctly and i am not currently tracking removals of children in prefab instances so it will come back